feat(invitation): add admin group invitation list functionality
Implement a new feature to allow users to retrieve a list of group invitations they have
received as a group admin. This includes the addition of a new method in the SysGroupInvitationService
interface and its implementation in SysGroupInvitationServiceImpl. Also, a new VO
(GroupAdminInvitationVo) has been created to store this invitation data which includes details
such as group name, inviter's information, and invitation status. Mapper XML and Java Mapper
interface modifications were made to support the new feature.